+++ ta4j-core/src/test/java/org/ta4j/core/XlsTestsUtils.java
@@ -31,8 +31,11 @@ import java.time.Instant;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Date;
+import java.util.Iterator;
 import java.util.List;
+import java.util.zip.DataFormatException;
 
 public class XlsTestsUtils {
 
@@ -43,15 +46,24 @@ public class XlsTestsUtils {
         return sheet;
     }
 
-    public static void setParamValue(Sheet sheet, int paramIndex, double value) {
-        // first param is in B3 cell
-        // second param is in B4 cell
-        Row row = sheet.getRow(2 + paramIndex);
-        Cell cell = row.getCell(1);
-        cell.setCellValue(value);
+    public static void setParams(Sheet sheet, Decimal... params) throws DataFormatException {
+        // the parameters follow a parameters header row with the first cell containing "Param"
+        FormulaEvaluator evaluator = sheet.getWorkbook().getCreationHelper().createFormulaEvaluator();
+        Iterator<Row> iterator = sheet.rowIterator();
+        while (iterator.hasNext()) {
+            Row row = iterator.next();
+            if (evaluator.evaluate(row.getCell(0)).formatAsString().contains("Param")) {
+                Arrays.asList(params)
+                    .stream()
+                    .mapToDouble(Decimal::doubleValue)
+                    .forEach(d -> iterator.next().getCell(1).setCellValue(d));
+                return;
+            }            
+        }
+        throw new DataFormatException("\"Param\" header row not found");
     }
 
-    public static TimeSeries readTimeSeries(Sheet sheet) {        
+    public static TimeSeries readTimeSeries(Sheet sheet) throws DataFormatException {        
         TimeSeries series = new BaseTimeSeries();
         FormulaEvaluator evaluator = sheet.getWorkbook().getCreationHelper().createFormulaEvaluator();
         Duration weekDuration = Duration.ofDays(7);
@@ -86,41 +98,33 @@ public class XlsTestsUtils {
         return values;
     }
     
-    public static List<Row> readDataAfterHeader(Sheet sheet) {
-        // header row contains "Date" in the first cell
-        // header row may appear anywhere in the sheet
-        // data starts after the header row
-        List<Row> rows = new ArrayList<Row>();
-        boolean noHeader = true;
+    public static List<Row> readDataAfterHeader(Sheet sheet) throws DataFormatException {
+        // the data follow a data header row with the first cell containing "Date"
         FormulaEvaluator evaluator = sheet.getWorkbook().getCreationHelper().createFormulaEvaluator();
-        for (int i = 0; i <= sheet.getLastRowNum(); i++) {
-            Row row = sheet.getRow(i);
-            if (row == null) {
-                continue;
+        Iterator<Row> iterator = sheet.rowIterator();
+        boolean noHeader = true;
+        List<Row> rows = new ArrayList<Row>();
+        while (iterator.hasNext()) {
+            Row row = iterator.next();
+            if (noHeader == false) {
+                rows.add(row);
             }
-            Cell cell = row.getCell(0);
-            CellValue value = evaluator.evaluate(cell);
-            if (value.formatAsString().contains("Date")) {
+            if (noHeader && evaluator.evaluate(row.getCell(0)).formatAsString().contains("Date")) {
                 noHeader = false;
-                continue;
             }
-            if (noHeader) {
-                continue;
-            }
-            rows.add(row);
+        }
+        if (noHeader) {
+            throw new DataFormatException("\"Date\" header row not found");
         }
         return rows;
-            
     }
 
-    public static void testXlsIndicator(Class testClass, String xlsFileName, int valueColumnIdx, IndicatorFactory indicatorFactory, Double... params) throws Exception {
+    public static void testXlsIndicator(Class testClass, String xlsFileName, int valueColumnIdx, IndicatorFactory indicatorFactory, Decimal... params) throws Exception {
         // read time series from xls
         Sheet sheet = getDataSheet(testClass, xlsFileName);
         TimeSeries inputSeries = readTimeSeries(sheet);
         // compute and read expected values from xls
-        for (int i = 0; i < params.length; i++) {
-            setParamValue(sheet, i, params[i]);
-        }
+        setParams(sheet, params);
         List<Decimal> expectedValues = readValues(sheet, valueColumnIdx);
         // create indicator using time series
         Indicator<Decimal> actualIndicator = indicatorFactory.createIndicator(inputSeries);
@@ -128,10 +132,10 @@ public class XlsTestsUtils {
         TATestsUtils.assertValuesEquals(actualIndicator, expectedValues);
     }
     public static <T> void testXlsIndicator(Class testClass, String xlsFileName, int valueColumnIdx, IndicatorFactory indicatorFactory, T... params) throws Exception {
-        Double[] doubleParams = new Double[params.length];
+        Decimal[] decimalParams = new Decimal[params.length];
         for (int i = 0; i < params.length; i++) {
-            doubleParams[i] = Double.valueOf(params[i].toString());
+            decimalParams[i] = Decimal.valueOf(params[i].toString());
         }
-        testXlsIndicator(testClass, xlsFileName, valueColumnIdx, indicatorFactory, doubleParams);
+        testXlsIndicator(testClass, xlsFileName, valueColumnIdx, indicatorFactory, decimalParams);
     }
 }